Rewrite main.js
